Added a first-pass dungeon generation algorithm.  Still not perfect.

+++ core/src/rml/dungeon/Entity.java
@@ -2,8 +2,8 @@ package rml.dungeon;
 
 class Entity {
 //    int EID;
-    float x, y;
-    float theta;
+    double x, y;
+    double theta;
     int txtID = -1;
     //int bbr_x = -1, bbr_y = -1;
     float radius = -1;
